Move shared Rich console to olmo_eval.common.console to break import cycle Co-Authored-By: Claude Opus 4.7 <noreply@anthropic.com>
