+++ src/olmo_eval/cli/utils.py
@@ -6,7 +6,8 @@ from datetime import datetime
 from typing import TYPE_CHECKING, Any
 
 import click
-from rich.console import Console
+
+from olmo_eval.common.console import console
 
 if TYPE_CHECKING:
     from olmo_eval.evals.external.base import ExternalEval
@@ -14,9 +15,6 @@ if TYPE_CHECKING:
     from olmo_eval.harness import HarnessConfig
     from olmo_eval.inference.providers.config import ProviderConfig
     from olmo_eval.launch.beaker.launcher import BeakerJobConfig
-
-
-console = Console(force_terminal=True, width=120)
 
 
 @dataclass
